Parse games endpoint and store in database
A random (last in chron data) value for the start time is stored
in the database, keyed by [season, day]. Values from gamma or the
coffee cup are currently discarded.

Data is currently unused, but is intended for use for annotation
of idol board data, and for improved navigation.

Currently using cached data containing all of chron/v1/games --
possible that this could pull from chron, and probably should if
the idol board comes back and we want dynamic data capture.

Sled DB is overwritten on every liftoff, which is costly when no
new updates are occuring.